requirements.txt
# Core dependencies
streamlit>=1.22.0
python-dotenv>=0.21.0

# Data handling and analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.4

# Visualization
matplotlib>=3.5.0
plotly>=5.15.0
seaborn>=0.12.0

# AI and Machine Learning
google-generativeai>=0.3.0
langchain>=0.0.267
langchain-google-genai>=0.0.4

# Web and API
requests>=2.28.1
flask>=2.0.1
duckduckgo-search>=2.9.5

# Data processing
beautifulsoup4>=4.12.0
html5lib>=1.1
lxml>=4.9.0

# Financial analysis
ta>=0.10.0  # Technical Analysis library
finplot>=1.9.0
pandas_datareader>=0.10.0

# Utilities
tqdm>=4.65.0
pytz>=2023.3
python-dateutil>=2.8.2

# State management
streamlit-extras>=0.3.0